+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 - [fuji] fix: should open the GUI if the `About GUI` is opened by `Modules Inspection GUI`.
 - [core] feature: attach detailed exception message for error log.
+- [fuji] feature: new `/fuji inspect jobs` command, to list all `registered jobs` and their `fire dates`.